ecommerce/requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
exceptiongroup==1.2.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
sentry-sdk==2.7.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
